+++ ClientApp/src/app/jobs/job-items-list-view/job-items-list-view.component.html
@@ -148,10 +148,11 @@
               <th colspan="6">
                 <div class="d-flex justify-content-start">
                   <button
-                    class="btn btn-falcon-primary btn-sm"
+                    class="btn btn-link btn-sm"
                     (click)="addLineItemClicked(true)"
                   >
-                    <i class="fa fa-plus"></i> Add product
+                    + Add product
+                    <!-- <i class="fa fa-plus"></i> Add product -->
                   </button>
                 </div>
               </th>
