+++ ClientApp/src/app/jobs/job-visits-view/job-visits-view.component.html
@@ -1,0 +1,1 @@
+<p>job-visits-view works!</p>
