+++ ClientApp/src/app/jobs/job-visits-view/job-visits-view.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input, OnInit } from "@angular/core";
+
+@Component({
+  selector: "app-job-visits-view",
+  templateUrl: "./job-visits-view.component.html",
+  styleUrls: ["./job-visits-view.component.scss"],
+})
+export class JobVisitsViewComponent implements OnInit {
+  @Input() jobId: number;
+  constructor() {}
+
+  ngOnInit(): void {}
+}
