+++ ClientApp/src/app/shared-ui-components/shared.data.service.ts
@@ -18,6 +18,38 @@ export class SharedDataService {
     private __customerServiceProxy: CustomersServiceProxy
   ) {}
 
+  get filterButtonsGroup() {
+    return [
+      { groupName: "*" },
+      { groupName: "A", checked: true },
+      { groupName: "B" },
+      { groupName: "C" },
+      { groupName: "D" },
+      { groupName: "E" },
+      { groupName: "F" },
+      { groupName: "G" },
+      { groupName: "H" },
+      { groupName: "I" },
+      { groupName: "J" },
+      { groupName: "K" },
+      { groupName: "L" },
+      { groupName: "M" },
+      { groupName: "N" },
+      { groupName: "O" },
+      { groupName: "P" },
+      { groupName: "Q" },
+      { groupName: "R" },
+      { groupName: "S" },
+      { groupName: "T" },
+      { groupName: "U" },
+      { groupName: "V" },
+      { groupName: "W" },
+      { groupName: "X" },
+      { groupName: "Y" },
+      { groupName: "Z" },
+    ];
+  }
+
   showCustomerAddressModal(
     customer: number | CustomerDto
   ): Promise<AddressDto> {
